+++ tests/Support/AuthorRule.php
@@ -13,7 +13,7 @@ use Yiisoft\Rbac\RuleInterface;
  */
 final class AuthorRule implements RuleInterface
 {
-    public function execute(string $userId, Item $item, RuleContext $ruleContext): bool
+    public function execute(?string $userId, Item $item, RuleContext $ruleContext): bool
     {
         return $ruleContext->getParameterValue('authorID') === $userId;
     }